rename: 파일명 interestChannels.js로 변경 (API 경로와 통일)
- API 경로(`/users/:userId/interest-channels`)와 명칭을 일치시키기 위해 파일명을 복수형으로 변경
